move complex logic to orchestration functions
